move lighthouse code to shared folder
