+++ src/lib/lighthouse/index.js
