+++ src/lib/lighthouse/mapping.js
